Consolidate public view notice
