+++ app/src/router.ts
@@ -4,7 +4,6 @@ import AcceptInviteRoute from '@/routes/accept-invite.vue';
 import LoginRoute from '@/routes/login/login.vue';
 import LogoutRoute from '@/routes/logout.vue';
 import PrivateNotFoundRoute from '@/routes/private-not-found.vue';
-import PublicNotice from '@/routes/public-notice.vue';
 import ResetPasswordRoute from '@/routes/reset-password/reset-password.vue';
 import ShareRoute from '@/routes/shared/shared.vue';
 import TFASetup from '@/routes/tfa-setup.vue';
@@ -22,15 +21,7 @@ export const defaultRoutes: RouteRecordRaw[] = [
 	{
 		name: 'login',
 		path: '/login',
-		components: {
-			default: LoginRoute,
-			notice: PublicNotice,
-		},
-		props: {
-			notice: (route) => ({
-				logoutReason: route.query.reason,
-			}),
-		},
+		component: LoginRoute,
 		meta: {
 			public: true,
 			view: 'public',
@@ -39,7 +30,7 @@ export const defaultRoutes: RouteRecordRaw[] = [
 	{
 		name: 'reset-password',
 		path: '/reset-password',
-		components: { default: ResetPasswordRoute, notice: PublicNotice },
+		component: ResetPasswordRoute,
 		meta: {
 			public: true,
 			view: 'public',
@@ -51,6 +42,7 @@ export const defaultRoutes: RouteRecordRaw[] = [
 		component: AcceptInviteRoute,
 		meta: {
 			public: true,
+			view: 'public',
 		},
 	},
 	{
